@@ -13,25 +13,25 @@ dependencies:
 
 
   cupertino_icons: ^1.0.8
-  table_calendar: ^3.1.2
+  table_calendar: ^3.1.3
   path_provider: ^2.1.4
   intl: ^0.19.0
   provider: ^6.1.2
   google_fonts: ^6.2.1
   animations: ^2.0.11
-  shared_preferences: ^2.3.1
+  shared_preferences: ^2.5.2
   carousel_slider: ^5.0.0
-  flutter_local_notifications: ^17.2.3
-  photo_manager_image_provider: ^2.1.1
-  photo_manager: ^3.0.0
-  camera: ^0.11.0+1
+  flutter_local_notifications: ^18.0.1
+  photo_manager_image_provider: ^2.2.0
+  photo_manager: ^3.6.4
+  camera: ^0.11.1
   flutter_staggered_animations: ^1.1.1
-  simple_animations: ^5.0.2
-  webview_flutter: ^4.8.0
+  simple_animations: ^5.1.0
+  webview_flutter: ^4.10.0
   csshadow: ^0.0.2
-  flutter_localization: ^0.2.1
+  flutter_localization: ^0.3.1
   permission_handler: ^10.4.0
-  hive: ^2.0.0
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
   uuid: ^4.4.2
   audioplayers: ^6.1.0
