@@ -1,7 +1,8 @@
 import 'dart:collection';
 import 'package:flutter/material.dart';
 
-import '../settings_model/social_model.dart';
+import '../../../models/settings_model/settings_model/social_model.dart';
+
 
 class LibList {
   final List<SocialModel> _libList = [
