@@ -1,25 +1,33 @@
 import 'dart:async';
 import 'dart:collection';
-import 'package:noti/providers/settings_provider.dart';
+import 'package:noti/providers/settings_provider/settings_provider.dart';
 import 'package:flutter/widgets.dart';
+import 'package:noti/providers/task_provider/task_search_provider.dart';
 import 'package:table_calendar/table_calendar.dart';
-import '../database/database_helper.dart';
-import '../models/db_model/task.dart';
-import '../utils/notifications/notifications_helper.dart';
-import '../utils/prefs/prefs.dart';
+import '../../database/database_helper.dart';
+import '../../models/db_model/task.dart';
+import '../../utils/notifications/notifications_helper.dart';
+import '../../utils/prefs/prefs.dart';
+
 
 class TaskProvider extends ChangeNotifier {
 
   final DatabaseHelper _dbHelper = DatabaseHelper.databaseHelper;
 
+  SettingsProvider settings;
+  TaskSearchProvider searchProvider;
+
+  TaskProvider(this.settings, this.searchProvider) {
+    initTask();
+  }
 
   int selectedIndex = 0;
   DateTime focDay =
   DateTime(DateTime.now().year, DateTime.now().month, DateTime.now().day,DateTime.now().hour,DateTime.now().minute);
   DateTime selDay =
   DateTime(DateTime.now().year, DateTime.now().month, DateTime.now().day,DateTime.now().hour,DateTime.now().minute);
-  DateTime? _rangeStart;
-  DateTime? _rangeEnd;
+  // DateTime? _rangeStart;
+  // DateTime? _rangeEnd;
   RangeSelectionMode rangeSelectionMode = RangeSelectionMode.toggledOff;
   Map<DateTime, List<Task>> tasks = {};
   List<Task> _taskList = [];
@@ -31,18 +39,23 @@ class TaskProvider extends ChangeNotifier {
 
   final Prefs _prefs = Prefs();
 
-  SettingsProvider settings;
-
-  TaskProvider(this.settings) {
-    initTask();
-  }
-
   Future<void> initTask() async {
-    _taskList = _dbHelper.getAllTasks();
+    //_taskList = _dbHelper.getAllTasks();
     selDay = focDay;
     loadCalendarFormat();
-    getSettingsValuesForTask().whenComplete(() => getTaskDbList());
+    getSettingsValuesForTask().whenComplete((){
+      getTaskDbList();
+      getTasksBySearchOptions();
+    });
     notifyListeners();
+  }
+  List<Task> _taskListByKeyword = [];
+  UnmodifiableListView<Task> get taskListByKeyword {
+    return UnmodifiableListView(_taskListByKeyword);
+  }
+
+  int get taskListByKeywordCounter {
+    return _taskListByKeyword.length;
   }
 
   UnmodifiableListView<Task> get taskList {
@@ -74,8 +87,8 @@ class TaskProvider extends ChangeNotifier {
     if (!isSameDay(selDay, selectedDay)) {
       selDay = selectedDay;
       focDay = DateTime(DateTime.now().year, DateTime.now().month, DateTime.now().day,DateTime.now().hour,DateTime.now().minute);
-      _rangeStart = null;
-      _rangeEnd = null;
+      // _rangeStart = null;
+      // _rangeEnd = null;
       rangeSelectionMode = RangeSelectionMode.toggledOff;
     }
     _taskList = getCalendarValues(selDay);
@@ -86,8 +99,8 @@ class TaskProvider extends ChangeNotifier {
     if (!isSameDay(selDay, selectedDay)) {
       selDay = selectedDay;
       focDay = focusedDay;
-      _rangeStart = null;
-      _rangeEnd = null;
+      // _rangeStart = null;
+      // _rangeEnd = null;
       rangeSelectionMode = RangeSelectionMode.toggledOff;
     }
     _taskList = getCalendarValues(selDay);
@@ -107,7 +120,15 @@ class TaskProvider extends ChangeNotifier {
       tasks[addDate] = [task];
     }
 
-    tasks[addDate]!.sort((a, b) => a.date.compareTo(b.date));
+    // tasks[addDate]!.sort((a, b) => a.date.compareTo(b.date));
+    tasks.forEach((key, value) {
+      value.sort((a, b) {
+        if (a.isTaskDone == b.isTaskDone) {
+          return a.date.compareTo(b.date); // Jeśli oba są w tej samej grupie (false/true), sortuj po dacie
+        }
+        return a.isTaskDone ? 1 : -1; // Przesuń `true` na dół, `false` na górę
+      });
+    });
 
     _taskList = getCalendarValues(focDay);
 
@@ -128,8 +149,14 @@ class TaskProvider extends ChangeNotifier {
       refreshNotification(task);
     }
 
+    // Sortowanie: Najpierw niewykonane, potem wykonane, w każdej grupie sortujemy po dacie.
     tasks.forEach((key, value) {
-      value.sort((a, b) => a.date.compareTo(b.date));
+      value.sort((a, b) {
+        if (a.isTaskDone == b.isTaskDone) {
+          return a.date.compareTo(b.date); // Jeśli oba są w tej samej grupie (false/true), sortuj po dacie
+        }
+        return a.isTaskDone ? 1 : -1; // Przesuń `true` na dół, `false` na górę
+      });
     });
 
     _taskList = getCalendarValues(focDay);
@@ -223,7 +250,8 @@ class TaskProvider extends ChangeNotifier {
   }
 
   void refreshNotification(Task task) {
-    print("NOTIFICATIONS DATES ${task.date}");
+    //todo: remove prints:
+    //print("NOTIFICATIONS DATES ${task.date}");
 
     if (settings.isNotification) {
       if (task.isTaskDone) {
@@ -293,6 +321,80 @@ class TaskProvider extends ChangeNotifier {
 
     notifyListeners();
     return tempList;
+  }
+
+  Future<List<Task>> getTasksBySearchOptions() async {
+    List<Task> list = _dbHelper.getAllTasks();
+
+
+    if (searchProvider.keyword.isEmpty &&
+        searchProvider.startDate == searchProvider.endDate &&
+        searchProvider.priority == -1 &&
+        !searchProvider.isDone) {
+      _taskListByKeyword = _dbHelper.getAllTasks();
+    } else {
+      _taskListByKeyword = list; // Zaczynamy od pełnej listy
+
+      // Filtrowanie po słowie kluczowym
+      if (searchProvider.keyword.isNotEmpty) {
+        _taskListByKeyword = _taskListByKeyword.where((task) {
+          return task.title.toLowerCase().contains(searchProvider.keyword.toLowerCase()) ||
+              task.description.toLowerCase().contains(searchProvider.keyword.toLowerCase());
+        }).toList();
+      }
+
+      // Filtrowanie po dacie
+      if (searchProvider.startDate.isBefore(searchProvider.endDate) &&
+          searchProvider.endDate.isAfter(searchProvider.startDate)) {
+        _taskListByKeyword = _taskListByKeyword.where((task) {
+          return task.date.isAfter(searchProvider.startDate) &&
+              task.date.isBefore(searchProvider.endDate);
+        }).toList();
+      }
+
+      // Filtrowanie po statusie (isTaskDone) – tylko jeśli użytkownik zaznaczył opcję
+      if (searchProvider.isDone) {
+        _taskListByKeyword = _taskListByKeyword.where((task) => task.isTaskDone).toList();
+      }
+
+      // Filtrowanie po priorytecie – tylko jeśli użytkownik wybrał inny niż -1
+      if (searchProvider.priority != -1) {
+        _taskListByKeyword = _taskListByKeyword.where((task) => task.priority == searchProvider.priority).toList();
+      }
+
+      notifyListeners();
+    }
+
+    // Sortowanie: Najpierw niewykonane, potem wykonane, w każdej grupie sortujemy po dacie
+    _taskListByKeyword.sort((a, b) {
+      if (a.isTaskDone == b.isTaskDone) {
+        return a.date.compareTo(b.date); // Jeśli oba są w tej samej grupie (false/true), sortuj po dacie
+      }
+      return a.isTaskDone ? 1 : -1; // Przesuń `true` na dół, `false` na górę
+    });
+
+    notifyListeners();
+    return _taskListByKeyword;
+  }
+
+
+  void resetTaskSearch() {
+    searchProvider.resetSearchFilters();
+    _taskListByKeyword = _dbHelper.getAllTasks();
+    notifyListeners();
+  }
+
+  void deleteSelectedTasks()async {
+    await getTasksBySearchOptions().then((tasks){
+      for(Task task in tasks){
+
+        //todo:
+        print("SELECTED NOTES TO DELETE ${task.title}");
+        deleteTask(task);
+      }
+    });
+    resetTaskSearch();
+    notifyListeners();
   }
 
   Future<List<Task>> deleteAllTasks() async{
