@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_staggered_animations/flutter_staggered_animations.dart';
-import '../../providers/settings_provider.dart';
-import '../../utils/constans/durations.dart';
+import '../../providers/settings_provider/settings_provider.dart';
+import '../../utils/constants/const_values.dart';
 import '../../utils/dimensions/screen_type.dart';
 import '../../utils/dimensions/size_info.dart';
 import '../../widgets/cards/theme_card.dart';
