@@ -2,8 +2,8 @@ import 'package:carousel_slider/carousel_slider.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:simple_animations/simple_animations.dart';
-import '../../providers/settings_provider.dart';
-import '../../utils/constans/durations.dart';
+import '../../providers/settings_provider/settings_provider.dart';
+import '../../utils/constants/const_values.dart';
 import '../../utils/dimensions/size_info.dart';
 
 class Carousel extends StatelessWidget {
