@@ -3,9 +3,9 @@ import 'package:camera/camera.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_staggered_animations/flutter_staggered_animations.dart';
-import '../../providers/gallery_image_provider.dart';
+import '../../providers/image_provider/gallery_image_provider.dart';
 import '../../utils/dimensions/size_info.dart';
-import '../buttons/icon_button.dart';
+import '../buttons/text_icon_button.dart';
 import '../cameraLoader/camera_loader.dart';
 import '../cards/image_card.dart';
 
@@ -19,16 +19,60 @@ class GallerySheet extends StatefulWidget {
 }
 
 class _GallerySheetState extends State<GallerySheet>
-    with SingleTickerProviderStateMixin {
+    with TickerProviderStateMixin {
   late TabController _tabController;
   int currentPage = 0;
   int? lastPage;
   bool expand = true;
 
+  late AnimationController _fadeController;
+  late Animation<double> _fadeAnimation;
+
   late CameraController _cameraController;
   bool _isInit = false;
   bool _isFlashOn = false;
+  bool _isCameraReverse = false;
 
+  final DraggableScrollableController _scrollableController = DraggableScrollableController();
+
+  bool isFullyCollapsed = false;
+
+  Future<void> switchCamera() async {
+
+    await _fadeController.forward().whenComplete(()async{
+      final cameras = await availableCameras();
+
+      CameraDescription? newCamera;
+      if (_isCameraReverse == false) {
+        newCamera = cameras.firstWhere(
+              (camera) => camera.lensDirection == CameraLensDirection.front,
+          orElse: () => cameras.first,
+        );
+      } else {
+        newCamera = cameras.firstWhere(
+              (camera) => camera.lensDirection == CameraLensDirection.back,
+          orElse: () => cameras.first,
+        );
+      }
+
+      if (newCamera != null) {
+        _cameraController.dispose();
+        _cameraController = CameraController(
+          newCamera,
+          ResolutionPreset.max,
+          enableAudio: false,
+        );
+
+        await _cameraController.initialize();
+
+        setState(() {
+          _isInit = true;
+          _isCameraReverse = !_isCameraReverse;
+        });
+      }
+    }).then((val)=>_fadeController.reverse());
+
+  }
 
 
   @override
@@ -41,7 +85,51 @@ class _GallerySheetState extends State<GallerySheet>
     });
     super.initState();
 
+    _fadeController = AnimationController(
+      vsync: this,
+      duration: const Duration(milliseconds: 250), // Szybka animacja
+    );
+
+    _fadeAnimation = Tween<double>(
+      begin: 1.0,
+      end: 0.0,
+    ).animate(CurvedAnimation(
+      parent: _fadeController,
+      curve: Curves.easeInOut,
+    ));
+
+    _scrollableController.addListener(() {
+      double currentSize = _scrollableController.size;
+
+      if (currentSize <= 0.2) {
+        isFullyCollapsed = true;
+      } else {
+        isFullyCollapsed = false;
+      }
+    });
+
     WidgetsBinding.instance.addPostFrameCallback((_) async {
+      final cameras = await availableCameras();
+      final initialCamera = cameras.firstWhere(
+            (camera) => camera.lensDirection == CameraLensDirection.back,
+        orElse: () => cameras.first,
+      );
+
+      _cameraController = CameraController(
+        initialCamera,
+        ResolutionPreset.max,
+        enableAudio: false,
+      );
+
+      await _cameraController.initialize();
+
+      setState(() {
+        _isInit = true;
+      });
+
+      await _cameraController.setFlashMode(FlashMode.off);
+    });
+   /* WidgetsBinding.instance.addPostFrameCallback((_) async {
       await availableCameras().then((value) {
         _cameraController = CameraController(value[0], ResolutionPreset.max,
             enableAudio: false);
@@ -53,7 +141,7 @@ class _GallerySheetState extends State<GallerySheet>
             });
       });
       await _cameraController.setFlashMode(FlashMode.off);
-    });
+    });*/
   }
 
   @override
@@ -67,6 +155,7 @@ class _GallerySheetState extends State<GallerySheet>
   Widget build(BuildContext context) {
     var tabTitleSize = SizeInfo.noteCardTitle;
     var bottomSheetBorderRadius = 20.0;
+
     return GestureDetector(
       onTap: () {
         Navigator.of(context).pop();
@@ -74,8 +163,9 @@ class _GallerySheetState extends State<GallerySheet>
       child: Container(
         color: Colors.transparent,
         child: DraggableScrollableSheet(
-            initialChildSize: .5,
-            minChildSize: .1,
+            controller: _scrollableController,
+            initialChildSize: .6,
+            minChildSize: .2,
             maxChildSize: .8,
             expand: expand,
             builder: (context, scrollController) {
@@ -100,6 +190,7 @@ class _GallerySheetState extends State<GallerySheet>
                               children: [
                                 NotificationListener<ScrollNotification>(
                                   onNotification: (ScrollNotification scroll) {
+
                                     if (scroll.metrics.pixels /
                                             scroll.metrics.maxScrollExtent >
                                         0.4) {
@@ -110,61 +201,91 @@ class _GallerySheetState extends State<GallerySheet>
                                     }
                                     return true;
                                   },
-                                  child: AnimationLimiter(
-                                    child: GridView.builder(
-                                      controller: scrollController,
-                                      physics: const BouncingScrollPhysics(
-                                        parent: AlwaysScrollableScrollPhysics(),
-                                      ),
-                                      shrinkWrap: true,
-                                      padding:  EdgeInsets.all(cameraPadding),
-                                      gridDelegate:
-                                          const SliverGridDelegateWithFixedCrossAxisCount(
-                                        crossAxisCount: 3,
-                                        mainAxisSpacing: 5.0,
-                                        crossAxisSpacing: 5.0,
-                                      ),
-                                      itemCount: imageProvider.imageListCounter,
-                                      itemBuilder: (context, index) {
-                                        if (imageProvider.imageList.isEmpty) {
-                                          return Center(
-                                            child: Text(
-                                              'Loading images',
-                                              style: txtStyle,
-                                            ),
-                                          );
-                                        } else {
-                                          final img =
-                                              imageProvider.imageList[index];
 
-                                          return AnimationConfiguration
-                                              .staggeredGrid(
-                                            columnCount: 3,
-                                            position: index,
-                                            duration: const Duration(
-                                                milliseconds: 374),
-                                            child: ScaleAnimation(
-                                              scale: 0.5,
-                                              child: FadeInAnimation(
-                                                child: ImageAssetCard(
-                                                  img: img,
-                                                  onTap: () async {
-                                                    img.file.then((value) {
-                                                      if (value != null) {
-                                                        return widget.pickImage(
-                                                            value
-                                                                .readAsBytesSync());
-                                                      }
-                                                    });
-                                                    Navigator.of(context).pop();
-                                                  },
-                                                ),
-                                              ),
+                                  child: Column(
+                                    mainAxisAlignment:MainAxisAlignment.start,
+                                    crossAxisAlignment: CrossAxisAlignment.start,
+                                    children: [
+                                      Expanded(
+                                        child: AnimationLimiter(
+                                          child: GridView.builder(
+                                            controller: scrollController,
+                                            physics: const BouncingScrollPhysics(
+                                              parent: AlwaysScrollableScrollPhysics(),
                                             ),
-                                          );
-                                        }
-                                      },
-                                    ),
+                                            shrinkWrap: true,
+                                            padding:  EdgeInsets.all(cameraPadding),
+                                            gridDelegate:
+                                                const SliverGridDelegateWithFixedCrossAxisCount(
+                                              crossAxisCount: 3,
+                                              mainAxisSpacing: 5.0,
+                                              crossAxisSpacing: 5.0,
+                                            ),
+                                            itemCount: imageProvider.imageListCounter,
+                                            itemBuilder: (context, index) {
+                                              if (imageProvider.imageList.isEmpty) {
+                                                return Center(
+                                                  child: Text(
+                                                    'Loading images',
+                                                    style: txtStyle,
+                                                  ),
+                                                );
+                                              } else {
+                                                final img =
+                                                    imageProvider.imageList[index];
+
+                                                return AnimationConfiguration
+                                                    .staggeredGrid(
+                                                  columnCount: 3,
+                                                  position: index,
+                                                  duration: const Duration(
+                                                      milliseconds: 374),
+                                                  child: ScaleAnimation(
+                                                    scale: 0.5,
+                                                    child: FadeInAnimation(
+                                                      child: ImageAssetCard(
+                                                        img: img,
+                                                        onTap: () async {
+                                                          img.file.then((value) {
+                                                            if (value != null) {
+                                                              return widget.pickImage(
+                                                                  value
+                                                                      .readAsBytesSync());
+                                                            }
+                                                          });
+                                                          Navigator.of(context).pop();
+                                                        },
+                                                      ),
+                                                    ),
+                                                  ),
+                                                );
+                                              }
+                                            },
+                                          ),
+                                        ),
+                                      ),
+                                      Visibility(
+                                        visible: !isFullyCollapsed,
+                                        child: Container(
+                                          padding: EdgeInsets.all(cameraPadding),
+                                          height: 60,
+                                          child: ListView.builder(
+                                            itemCount: imageProvider.albumsListCounter,
+                                            shrinkWrap: true,
+                                            padding: EdgeInsets.symmetric(horizontal:cameraPadding),
+                                            itemBuilder: (context, index){
+                                              var album = imageProvider.albumsList[index];
+                                              return TextButton(
+                                                  onPressed: (){
+                                                    imageProvider.onAlbumChoose(index);
+                                                  }, child: Text(album.name,style: Theme.of(context).textTheme.bodyMedium!.copyWith(
+                                                  fontSize: tabTitleSize,color: imageProvider.selectedAlbum == index ? Theme.of(context).indicatorColor : Theme.of(context).textTheme.bodyMedium!.color
+                                              ),)
+                                              );
+                                            },scrollDirection: Axis.horizontal,),
+                                        ),
+                                      ),
+                                    ],
                                   ),
                                 ),
                                 _isInit
@@ -176,27 +297,17 @@ class _GallerySheetState extends State<GallerySheet>
                                           spacing: cameraPadding,
                                           children: [
                                             Expanded(
-                                              child: RotatedBox(
-                                                  quarterTurns: 1,
-                                                  child: ClipRRect(
-                                                    borderRadius: BorderRadius.all(Radius.circular(bottomSheetBorderRadius/2)),
-                                                    child: CameraPreview(
-                                                      _cameraController,
-                                                    ),
-                                                  )
-                                                  // CameraScreen(
-                                                  //                                 key: widget.key,
-                                                  //                                 controller: _cameraController,
-                                                  //                                 scrollController: scrollController,
-                                                  //                                 onPhoto: () async {
-                                                  // final res = await _cameraController.takePicture();
-                                                  // final value = await res.readAsBytes();
-                                                  // imageProvider.addTakenPictureToGallery(value);
-                                                  // widget.pickImage(value);
-                                                  //  Navigator.pop(context);
-                                                  //                                 },
-                                                  //                               ),
-                                                  ),
+                                              child: FadeTransition(
+                                                opacity: _fadeAnimation,
+                                                child: RotatedBox(
+                                                    quarterTurns: _isCameraReverse ? 3 : 1,
+                                                    child: ClipRRect(
+                                                      borderRadius: BorderRadius.all(Radius.circular(bottomSheetBorderRadius/2)),
+                                                      child: CameraPreview(
+                                                        _cameraController,
+                                                      ),
+                                                    )),
+                                              )
                                             ),
                                             Row(
                                               mainAxisAlignment: MainAxisAlignment.spaceEvenly,
@@ -215,6 +326,18 @@ class _GallerySheetState extends State<GallerySheet>
                                                   widget.pickImage(value);
                                                    Navigator.pop(context);
                                                    },
+                                                ),
+                                                IconButtonWithText(
+                                                  value: _isCameraReverse,
+                                                  iconData: Icons.cameraswitch,
+                                                  iconName: 'camera switch',
+                                                  padding: EdgeInsets.all(cameraPadding),
+                                                  onChanged: (value) async {
+                                                    // setState(() {
+                                                    //   _isCameraReverse = !_isCameraReverse;
+                                                    // });
+                                                     await switchCamera();
+                                                  },
                                                 ),
                                                 IconButtonWithText(
                                                   value: _isFlashOn,
