@@ -33,6 +33,7 @@ class TaskDatePickerDial extends StatefulWidget {
 }
 
 class _TaskDatePickerDialState extends State<TaskDatePickerDial> {
+
   late DateTime focDay;
   late DateTime selDay;
 
@@ -69,12 +70,6 @@ class _TaskDatePickerDialState extends State<TaskDatePickerDial> {
       }
     }
     return widget.scopeDatesList;
-  }
-
-  void onExpanded(){
-    setState(() {
-      isExpanded = !isExpanded;
-    });
   }
 
   void onScopeDateSelected(){
@@ -138,6 +133,12 @@ class _TaskDatePickerDialState extends State<TaskDatePickerDial> {
     });
   }
 
+  void getFullMonth(DateTime focDay){
+
+    startDate = DateTime(focDay.year, focDay.month, 1);
+    endDate = DateTime(focDay.year, focDay.month + 1, 1).subtract(Duration(days: 1));
+
+  }
 
   List<DateTime> getCalendarDates(DateTime date) {
     return widget.scopeDatesList.where((item) {
@@ -180,6 +181,7 @@ class _TaskDatePickerDialState extends State<TaskDatePickerDial> {
               mainAxisSize: MainAxisSize.min,
               crossAxisAlignment: CrossAxisAlignment.center,
               children: [
+                //calendar header
                 Container(
                   width: MediaQuery.of(context).size.width,
                   height: textSize * 3,
@@ -223,7 +225,6 @@ class _TaskDatePickerDialState extends State<TaskDatePickerDial> {
                     ],
                   ),
                 ),
-            
                 //calendar
                 Padding(
                   padding: const EdgeInsets.symmetric(horizontal: 8.0,vertical: 0),
@@ -232,6 +233,8 @@ class _TaskDatePickerDialState extends State<TaskDatePickerDial> {
                     focDay: focDay,
                     selDay: selDay,
                     markerColor:markerColor,
+                    rangeStartDay: startDate,
+                    rangeEndDay: endDate,
                     onDaySelected: (selectedDay, focusedDay) {
 
                         setDialState(() {
@@ -322,7 +325,6 @@ class _TaskDatePickerDialState extends State<TaskDatePickerDial> {
                               children: [
             
                                 TextButton(
-            
                                     style: ButtonStyle(
                                         shape:WidgetStateProperty.all(
                                           RoundedRectangleBorder(
@@ -391,6 +393,50 @@ class _TaskDatePickerDialState extends State<TaskDatePickerDial> {
                                       fontSize:pickerSubtitle,
                                       color: !isDateScopeSelected ? selectedDateColor : baseColor,)
             
+                                        )
+                                      ]
+                                  ),
+                                )),
+                                Text(
+                                  'OR',
+                                  style: Theme.of(context).inputDecorationTheme.helperStyle!.copyWith(
+                                    color: baseColor,
+                                  ),
+                                ),
+                                TextButton(
+                                    style: ButtonStyle(
+                                        shape:WidgetStateProperty.all(
+                                          RoundedRectangleBorder(
+                                            borderRadius: BorderRadius.circular(10.0),
+                                          ),
+                                        ) ,
+                                        side: WidgetStateProperty.all(BorderSide(
+                                          width: 0.5,
+                                          color: baseColor!,
+                                        ))
+                                    ),
+                                    onPressed: (){
+
+                                      setState((){
+                                        getFullMonth(focDay);
+                                      });
+                                    }, child: RichText(
+                                  text: TextSpan(
+                                      text: "full month\n",
+                                      style: Theme.of(context).inputDecorationTheme.helperStyle!.copyWith(
+                                        color: !isDateScopeSelected ? selectedDateColor : baseColor,
+                                      ),
+                                      children: <TextSpan>[
+                                        TextSpan(
+                                            text: DateFormat('MMM yy').format(focDay),
+                                            style: Theme.of(context)
+                                                .textTheme
+                                                .headlineMedium!
+                                                .copyWith(
+                                              height: 1.5,
+                                              fontSize:pickerSubtitle,
+                                              color: !isDateScopeSelected ? selectedDateColor : baseColor,)
+
                                         )
                                       ]
                                   ),
