@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:simple_animations/animation_builder/play_animation_builder.dart';
-import '../../utils/constans/durations.dart';
+import '../../utils/constants/const_values.dart';
 import '../../utils/dimensions/size_info.dart';
 
 class SmallHeader extends StatelessWidget {
@@ -13,13 +13,13 @@ class SmallHeader extends StatelessWidget {
     return Padding(
       padding: EdgeInsets.symmetric(vertical: SizeInfo.verticalHeaderPadding),
       child: PlayAnimationBuilder(
-          tween: Tween<double>(begin: 0.9, end: 1.0),
+          tween: Tween<Offset>(begin: const Offset(-10.0, 0), end: Offset.zero),
           duration: headerDuration,
           delay: const Duration(milliseconds: 100),
           curve: Curves.easeOut,
         builder: (context, value, child){
-          return Transform.scale(
-            scale: value,
+          return Transform.translate(
+            offset: value,
             child: Text(
               title,
               style: Theme.of(context)
